requirements.txt
Django==3.2.16
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
django-model-utils==4.2.0
django-filter==22.1
drf-yasg==1.21.4
markdown==3.4.1
